Moved Hardware Functions from MouseBot to RobotIO
