+++ MicroMouse-2016.vcxproj
@@ -157,6 +157,7 @@
     <ClCompile Include="micromouse\Node.cpp" />
     <ClCompile Include="micromouse\Path.cpp" />
     <ClCompile Include="micromouse\RobotIO.cpp" />
+    <ClCompile Include="micromouse\IRSensor.cpp" />
     <ClCompile Include="micromouse\Vector.cpp" />
     <ClCompile Include="micromouse\VirtualMaze.cpp" />
   </ItemGroup>
@@ -169,6 +170,7 @@
     <ClInclude Include="micromouse\Node.h" />
     <ClInclude Include="micromouse\Path.h" />
     <ClInclude Include="micromouse\RobotIO.h" />
+    <ClInclude Include="micromouse\IRSensor.h" />
     <ClInclude Include="micromouse\Vector.h" />
     <ClInclude Include="micromouse\VirtualMaze.h" />
   </ItemGroup>
