+++ MicroMouse-2016.vcxproj.filters
@@ -43,6 +43,7 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="micromouse\VirtualMaze.cpp">
+    <ClCompile Include="micromouse\IRSensor.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -75,6 +76,7 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="micromouse\VirtualMaze.h">
+    <ClInclude Include="micromouse\IRSensor.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
