+++ micromouse/MouseBot.h
@@ -29,6 +29,10 @@ namespace Micromouse
 		void setPos(int x, int y);				// Sets the position to (x,y)
 		void setPos(PositionVector pos);		// Sets the position to pos
 
+		//Maps out the maze. If on Teensey, it uses the sensors and motors to map a physical maze.
+		//Otherwise, it generates a random, virtual maze and uses it to simulate mapping.
+		//The mouse will visit every cell.
+		//Currently, the mouse ends in a random cell. Later, this function should return the mouse to the start.
 		void mapMaze();
 
 		bool isClearForward();					// Returns true if there isn't a wall in front of the mouse. Uses a virtual maze for debugging on PC, otherwise it uses the bot's hardware.
@@ -43,17 +47,6 @@ namespace Micromouse
 		void rotateToFaceDirection(direction dir); // Rotates the mouse in place until it reaches the given direction.
 
 	private:
-		//TODO use actual pin numbers
-#ifdef __MK20DX256__ //this is the Teensy signature
-		//comment line below to test compile, remove #error when pins have been set
-#error define correct pin numbers 
-#endif
-		const int IR_LEFT_PIN = 1;
-		const int IR_RIGHT_PIN = 2;
-		const int IR_FRONT_LEFT_PIN = 3;
-		const int IR_FRONT_RIGHT_PIN = 4;
-
-		enum IRDirection { LEFT, RIGHT, FRONT_LEFT, FRONT_RIGHT };
 
 		void move(direction dir);
 		void followPath(Path* path);
@@ -68,12 +61,8 @@ namespace Micromouse
 		VirtualMaze* virtualMaze;
 		RobotIO robotIO;
 		stack<direction> movementHistory;
-		void initSensors();
 
 		PositionVector position = PositionVector(0,0);
 		direction facing = N;
-
-		IRSenor* IRSensors[4];
-
 	};
 }
