+++ micromouse/RobotIO.cpp
@@ -2,11 +2,26 @@
 
 namespace Micromouse
 {
-	// ## NOT YET IMPLEMENTED ## 
+	/**** CONSTRUCTORS ****/
+
 	RobotIO::RobotIO()
 	{
-
+		initSensors();
 	}
+
+	RobotIO::~RobotIO()
+	{
+		for (int i = 0; i < 4; i++)
+		{
+			delete IRSensors[i];
+		}
+	}
+
+
+
+
+
+	/**** SENSORS ****/
 
 	// ## NOT YET IMPLEMENTED ## 
 	bool RobotIO::isClearForward()
@@ -26,6 +41,12 @@ namespace Micromouse
 		return false;
 	}
 
+
+
+
+
+	/**** MOTORS ****/
+
 	// ## NOT YET IMPLEMENTED ## 
 	void RobotIO::moveForward()
 	{
@@ -42,5 +63,24 @@ namespace Micromouse
 	void RobotIO::rotateRight()
 	{
 
+	}
+
+
+
+
+
+	/**** INITIALIZATIONS ****/
+
+	void RobotIO::initSensors()
+	{
+		for (int i = 0; i < 4; i++)
+		{
+			delete IRSensors[i];
+		}
+
+		IRSensors[LEFT] = new IRSenor(IR_LEFT_PIN, 40, 300);
+		IRSensors[RIGHT] = new IRSenor(IR_RIGHT_PIN, 40, 300);
+		IRSensors[FRONT_LEFT] = new IRSenor(IR_FRONT_LEFT_PIN, 40, 300);
+		IRSensors[FRONT_RIGHT] = new IRSenor(IR_FRONT_RIGHT_PIN, 40, 300);
 	}
 }
